requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pikepdf==9.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
